backend/requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
astroid==3.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
dill==0.3.7
distro==1.8.0
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
isort==5.13.2
mccabe==0.7.0
openai==1.3.9
packaging==23.2
platformdirs==4.1.0
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tblib==3.0.0
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.1
urllib3==2.1.0

# Linting
pylint==3.0.3
pylint-django==2.5.5
black>=23.3.0
pylint-plugin-utils==0.8.2

# development
django-debug-toolbar>=2.2,<4.3
psycopg2-binary>=2.8,<2.10

# types
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.9.0

# django
Django==5.0
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5